requirements.txt
emoji==0.5.4
mysql-connector-python==8.0.18
python-telegram-bot==12.2.0
pandas==0.25.0
sqlalchemy==1.3.17
xlrd==1.2.0